SDK-228 - Send cellular network state data in lowercased like "2g","3g", etc SDK-222 - Use <site-id>.btttag.com instead of d.btttag.com while fetching config SDK-223 - Parse sample rate fields as float instead as integer while fetching config

+++ analytics/src/main/java/com/bluetriangle/analytics/dynamicconfig/model/BTTRemoteConfigurationMapper.kt
@@ -25,7 +25,7 @@ internal object BTTRemoteConfigurationMapper {
     private const val GROUPED_VIEW_SAMPLE_RATE = "groupedViewSampleRate"
 
     fun fromJson(remoteConfigJson: JSONObject): BTTRemoteConfiguration {
-        val networkSampleRate = remoteConfigJson.getIntOrNull(NETWORK_SAMPLE_RATE)?.div(100.0)
+        val networkSampleRate = remoteConfigJson.getDoubleOrNull(NETWORK_SAMPLE_RATE)?.div(100.0)
         val enableRemoteConfig = remoteConfigJson.getBooleanOrNull(ENABLE_REMOTE_CONFIG) == true
         val ignoreScreens = remoteConfigJson.getJsonArrayOrNull(IGNORE_SCREENS)?.let { array ->
             buildList {
